divide km-storage to multiple files
