+++ src/services/api/api-base.js
@@ -74,18 +74,17 @@ export function tryCastToApiError(error) {
   throw error
 }
 
-export function sanitizeUrlParam(value) {  
-  if (value instanceof(Date))  return value.toISOString();
-  if (value instanceof(Object)) return JSON.stringify(value);  
-  return value; 
+export function stringifyUrlParam(value) {  
+  if(value!==undefined)value=JSON.stringify(value);	
+  return value;
 }
 
 export function stringifyUrlParams(valueObj) {
   const returnObj = {};
 
   for (const [key, value] of Object.entries(valueObj)) {
-    if (value !== undefined){
-      returnObj[key] = sanitizeUrlParam(value);
+    if (isValid(value)){
+      returnObj[key] = stringifyUrlParam(value);
     }
   }
