+++ src/services/api/km-storage/KmDocuments.js
@@ -1,26 +1,18 @@
-import ApiBase, { tryCastToApiError, isValid, stringifyUrlParams } from './api-base';
+import ApiBase, { tryCastToApiError, isValid, stringifyUrlParams } from '../api-base';
 
 const  serviceUrls = { 
-  documentQueryUrl      (){ return "/api/v2013/documents/" },
+  documentQueryUrl      (){ return "/api/v2013/documents" },
   documentUrl           (identifier){ return `/api/v2013/documents/${encodeURIComponent(identifier)}` },
   documentInfoUrl       (identifier){ return `/api/v2013/documents/${encodeURIComponent(identifier)}/info` }, 
   validateUrl           (){ return "/api/v2013/documents/x/validate" },
-  draftUrl              (identifier){ return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/draft` },
   attachmentUrl         (identifier, filename) { return `/api/v2013/documents/${encodeURIComponent(identifier)}/attachments/${encodeURIComponent(filename)}` },
   temporaryAttachmentUrl(                    ) { return `/api/v2015/temporary-files` },
   persistAttachmentUrl  (identifier, guid    ) { return `/api/v2013/documents/${encodeURIComponent(identifier)}/attachments/persist-temporary/${encodeURIComponent(guid)}` },
   securityUrl           (identifier, operation){ return `/api/v2013/documents/${encodeURIComponent(identifier)}/securities/${encodeURIComponent(operation)}` },
-  draftSecurityUrl      (identifier, operation){ return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/draft/securities/${encodeURIComponent(operation)}` },
-  draftLockUrl          (identifier, lockID)   { return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/draft/locks/${encodeURIComponent(lockID)}` },
-  draftLockListUrl      (identifier)           { return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/draft/locks` },
   documentVersionListUrl(identifier)           { return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions` },
   documentVersionUrl    (identifier)           { return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/:revision` },
   documentVersionInfoUrl(identifier)           { return `/api/v2013/documents/${encodeURIComponent(identifier)}/versions/:revision/info` },
  
-  // exists  /api/v2013/documents/:uid/versions/:revision
-  // getinfo /api/v2013/documents/:uid/versions/:revision/info
-  // get     /api/v2013/documents/:uid/versions/:revision
-  // list    /api/v2013/documents/:uid/versions
 }
 
 const DefaultValues = {
@@ -28,19 +20,8 @@ const DefaultValues = {
   Accept : "application/json"
 }
 
-export default class KmStorageApi extends ApiBase 
-{
-  constructor(options) {
-    super(options);
-    
-    this.documents    = new KmDocumentsApi(options);
-    this.drafts       = new KmDraftsApi(options);
-    this.locks        = new KmLocksApi(options);
-    this.attachments  = new KmAttachmentsApi(options);
-  }
-}
 
-class KmDocumentsApi extends ApiBase
+export class KmDocumentsApi extends ApiBase
 {
   constructor(options) {
     super(options);
@@ -178,7 +159,7 @@ class KmDocumentsApi extends ApiBase
   }
 }
 
-class KmDocumentsVersionApi extends ApiBase
+export class KmDocumentsVersionApi extends ApiBase
 {
   constructor(options) {
     super(options);
@@ -217,7 +198,7 @@ class KmDocumentsVersionApi extends ApiBase
     const Realm = realm ? { Realm : realm } : {};  
     const headers =  { ...Realm }
 
-    return this.http.get(serviceUrls.documentVersionUrl(identifier) ,{ headers } )
+    return this.http.get(serviceUrls.documentVersionInfoUrl(identifier) ,{ headers } )
                     .then(res => res.data)
                     .catch(tryCastToApiError);
   }
@@ -235,209 +216,7 @@ class KmDocumentsVersionApi extends ApiBase
 
 }
 
-
-class KmDraftsApi extends ApiBase
-{
-  constructor(options) {
-
-    super(options);
-    this.self = this;
-    this.config.headers = {
-      Realm:   DefaultValues.Realm,
-      Accept : DefaultValues.Accept,
-    }
-  }
-
-  async query({realm, q, s, l, sk }={}){  
-    const params = stringifyUrlParams( {    
-      $filter:q, 
-      $orderby:s,
-      $top:l,
-      $skip:sk, 
-      collection : "mydraft"  
-    });
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    const data =  this.http.get(serviceUrls.documentQueryUrl(), { headers, params} )
-                           .then(res => res.data)
-                           .catch(tryCastToApiError);
-     
-    if(data?.Items?.length){
-      data.Items = data.Items.map(e=>{
-            if(e.workingDocumentBody){
-                e.body = e.workingDocumentBody;
-            }
-            return e;
-      });
-    }
-
-    return data;
-  }
-
-  async get(identifier,realm=null  ){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    const data =  this.http.get(serviceUrls.draftUrl(identifier), { headers })
-                           .then(res => res.data)
-                           .catch(tryCastToApiError);
-    
-    if(data.workingDocumentBody){
-        data.body = data.workingDocumentBody;
-    }
-
-    return data;
-  }
-
-  async exists( identifier, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-     return this.http.head(serviceUrls.draftUrl(identifier), { headers })
-                     .then(res => res.data)
-                     .catch(tryCastToApiError);
-  }
-
-  async put( identifier, body, {realm, schema}={}){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    if(!isValid(body)) throw Error(`invalid value for body`);  
-
-    const params = stringifyUrlParams( {schema}); 
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const ContentType = { 'Content-Type': 'application/json' };
-    const headers =  { ...Realm , ...ContentType}
-
-    return this.http.put(serviceUrls.draftUrl(identifier),  body , { headers, params })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-
-  async delete(identifier, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.delete(serviceUrls.draftUrl(identifier), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-  
-  async canCreate(identifier, {realm, schema, metadata }={}){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const params = stringifyUrlParams( {schema, metadata }); 
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.get(serviceUrls.draftSecurityUrl(identifier, 'create'), { headers, params })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-
-  async canUpdate(identifier, {realm, metadata}={}){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    const params = stringifyUrlParams( {metadata}); 
-    return this.http.get(serviceUrls.draftSecurityUrl(identifier, 'update'), { headers, params })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-
-  async canDelete(identifier, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.get(serviceUrls.draftSecurityUrl(identifier, 'delete'), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-}
-
-class KmLocksApi extends ApiBase
-{
-  constructor(options) {
-    super(options);
-    this.config.headers = {
-      Realm:   DefaultValues.Realm,
-      Accept : DefaultValues.Accept,
-    }
-  }
-
-  async query(identifier, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
- 
-    return this.http.get(serviceUrls.draftLockListUrl(identifier), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-
-  async get(identifier, lockID, realm=null ){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    if(!isValid(lockID)) throw Error(`invalid value for lockID`); 
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.get(serviceUrls.draftLockUrl(identifier, lockID), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-
-  async exists(identifier,lockID, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    if(!isValid(lockID)) throw Error(`invalid value for lockID`); 
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.head(serviceUrls.draftLockUrl(identifier, lockID), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }  
- 
-  async put(identifier,lockID, body, realm=null){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    if(!isValid(lockID)) throw Error(`invalid value for lockID`); 
-   
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-    
-    return this.http.put(serviceUrls.draftLockUrl(identifier,lockID), body, { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
- 
-  async delete(identifier,lockID, realm=null ){
-    if(!isValid(identifier)) throw Error(`invalid value for identifier`);  
-    if(!isValid(lockID)) throw Error(`invalid value for lockID`); 
-
-    const Realm = realm ? { Realm : realm } : {};  
-    const headers =  { ...Realm }
-
-    return this.http.delete(serviceUrls.draftLockUrl(identifier,lockID), { headers })
-                    .then(res => res.data)
-                    .catch(tryCastToApiError);
-  }
-}
-
-class KmAttachmentsApi extends ApiBase
+export class KmAttachmentsApi extends ApiBase
 {
   constructor(options) {
     super(options);
